Possible Future Research Directions.md:  Mention 'Begin…End' programming languages.
     As noted in the parenthetical statement accompanying the newly added bullet item, my opinion on these is
this:  it might be useful for a 'curly-brace' language to adapt/adopt into an optional annotation on a block's
closing delimiter the semantics of how (at least some of) these kinds of languages require their users to
explicitly mark statements closing blocks of code residing in programs written in these languages with the names
of those blocks of code which they close.  Though this violates the 'Don't Repeat Yourself' principle if
required, giving users the _option_ to do this would allow them to better convey their intent to other
programmers reading their code at the end of long-running blocks of it.  It might _also_ give compiler vendors
more information about program scopes so that they could have their products emit slightly better diagnostics
when detecting inconsistencies in scoping caused by user mistakes.  What block a closing delimiter ends would
then presumably then be deduced by the compiler or interpreter (as it is today in 'curly-brace' languages) if a
user does not specify this with said optional annotation.  (I won't make any prescription as to exactly what
form this optional annotation might take at this time.)